# apps/web/lib/api/openapi.json
{
  "components": {
    "schemas": {
      "AccountDeleteRequest": {
        "additionalProperties": false,
        "properties": {
          "email_confirmation": {
            "format": "email",
            "title": "Email Confirmation",
            "type": "string"
          }
        },
        "required": [
          "email_confirmation"
        ],
        "title": "AccountDeleteRequest",
        "type": "object"
      },
      "Body_uploadAvatar": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_uploadAvatar",
        "type": "object"
      },
      "EducationLevel": {
        "enum": [
          "lycee",
          "licence",
          "master",
          "doctorat",
          "autre"
        ],
        "title": "EducationLevel",
        "type": "string"
      },
      "ForgotPasswordRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "ForgotPasswordRequest",
        "type": "object"
      },
...
      "GoogleExchangeRequest": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 2048,
            "title": "Code",
            "type": "string"
          },
          "state": {
            "maxLength": 128,
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "code",
          "state"
        ],
        "title": "GoogleExchangeRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "OnboardingRequest": {
        "additionalProperties": false,
        "properties": {
          "age": {
            "maximum": 120.0,
            "minimum": 13.0,
            "title": "Age",
            "type": "integer"
          },
          "ai_familiarity": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Ai Familiarity",
            "type": "integer"
          },
          "education_level": {
            "$ref": "#/components/schemas/EducationLevel"
          }
        },
        "required": [
          "education_level",
          "age",
          "ai_familiarity"
        ],
        "title": "OnboardingRequest",
        "type": "object"
      },
      "PasswordChangeRequest": {
        "additionalProperties": false,
        "properties": {
          "current_password": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Password"
          },
          "new_password": {
            "maxLength": 128,
            "minLength": 8,
            "title": "New Password",
            "type": "string"
          }
        },
        "required": [
          "new_password"
        ],
        "title": "PasswordChangeRequest",
        "type": "object"
      },
      "ProfileUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "age": {
            "anyOf": [
              {
                "maximum": 120.0,
                "minimum": 13.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "ai_familiarity": {
            "anyOf": [
              {
                "maximum": 5.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Familiarity"
          },
          "education_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EducationLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "family_name": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Name"
          },
          "given_name": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Given Name"
          },
          "locale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale"
          },
          "pseudo": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pseudo"
          },
          "xai_audience": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/XaiAudience"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "ProfileUpdateRequest",
        "type": "object"
      },
      "RegisterRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
      "ResetPasswordRequest": {
        "additionalProperties": false,
        "properties": {
          "new_password": {
            "maxLength": 128,
            "minLength": 8,
            "title": "New Password",
            "type": "string"
          },
          "token": {
            "maxLength": 128,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "new_password"
        ],
        "title": "ResetPasswordRequest",
        "type": "object"
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserRead"
          }
        },
        "required": [
          "access_token",
          "expires_in",
          "user"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
      "UserRead": {
        "properties": {
          "age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "ai_familiarity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Familiarity"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "credits": {
            "title": "Credits",
            "type": "integer"
          },
          "education_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EducationLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "family_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Name"
          },
          "given_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Given Name"
          },
          "has_avatar": {
            "title": "Has Avatar",
            "type": "boolean"
          },
          "has_password": {
            "title": "Has Password",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "locale": {
            "title": "Locale",
            "type": "string"
          },
          "onboarding_completed": {
            "title": "Onboarding Completed",
            "type": "boolean"
          },
          "pseudo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pseudo"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "xai_audience": {
            "$ref": "#/components/schemas/XaiAudience"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "is_active",
          "pseudo",
          "given_name",
          "family_name",
          "locale",
          "education_level",
          "age",
          "ai_familiarity",
          "xai_audience",
          "credits",
          "has_password",
          "has_avatar",
          "onboarding_completed",
          "created_at"
        ],
        "title": "UserRead",
        "type": "object"
      },
      "UserRole": {
        "enum": [
          "user",
          "contributor",
          "admin"
        ],
        "title": "UserRole",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "XaiAudience": {
        "enum": [
          "novice",
          "intermediate",
          "expert"
        ],
        "title": "XaiAudience",
        "type": "string"
      }
    }
  },
  "info": {
    "description": "API d'IBIS-X v2 — plateforme d'accompagnement ML de bout en bout (sélection éthique de datasets, pipeline guidé, explicabilité).",
    "title": "IBIS-X API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/forgot-password": {
      "post": {
        "description": "Répond toujours 204 (pas d'énumération). Sans SMTP le lien est loggé (dev).",
        "operationId": "forgotPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Forgot Password",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/google/exchange": {
      "post": {
        "description": "Échange le code Google, valide l'id_token, émet NOS jetons (ADR-003).",
        "operationId": "googleExchange",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleExchangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Google Exchange",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/register": {
      "post": {
        "description": "Inscription email + mot de passe, auto-connexion (CDC §4.1).",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "operationId": "resetPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset Password",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/users/me": {
      "get": {
        "operationId": "getMe",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Me",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "updateMe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Me",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/me/avatar": {
      "get": {
        "operationId": "getMyAvatar",
        "responses": {
          "200": {
            "description": "Successful Response"
          }
        },
        "summary": "Get My Avatar",
        "tags": [
          "users"
        ]
      },
      "put": {
        "operationId": "uploadAvatar",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_uploadAvatar"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Avatar",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/me/delete": {
      "post": {
        "description": "Suppression de compte (confirmation par email saisi) — cascade projets/expériences.",
        "operationId": "deleteAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Account",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/me/onboarding": {
      "post": {
        "operationId": "completeOnboarding",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Onboarding",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/me/password": {
      "patch": {
        "operationId": "changePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change Password",
        "tags": [
          "users"
        ]
      }
    }
  }
}